+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - [beta] Add missing localizations.
 - [beta] Update the user guide.
 
+### Fixes
+
+- Fix an issue where scroll bars in the editor became invisible under certain theme and appearance combinations.
+
 
 
 6.2.0-rc (782)